backend/pyproject.toml
[project]
name = "autoresearch-cockpit"
version = "0.1.0"
description = "Control plane backend for karpathy/autoresearch"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "sse-starlette>=2.2",
    "gitpython>=3.1",
    "openai>=1.60",
    "anthropic>=0.42",
    "google-genai>=1.5",
    "httpx>=0.28",
    "cryptography>=44",
    "python-multipart>=0.0.18",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6",
    "httpx>=0.28",
    "aiosqlite>=0.20",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
